requirements.txt
requests==2.32.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
PyYAML==6.0.1
anthropic==0.25.0
supabase==2.3.5
httpx==0.27.2
python-dotenv==1.0.0
